# SMS-Onboarding-Unified/api-specification.yaml
openapi: 3.0.3
info:
  title: SMS Onboarding Portal API
  description: |
    RESTful API for the SMS Onboarding Portal that manages vessel onboarding, equipment documentation,
    critical parts management, and integration with the maintenance portal.
  version: 1.0.0
  contact:
    name: SMS Support Team
    email: support@sms-portal.com
  license:
    name: Proprietary
servers:
  - url: https://api.sms-onboarding.com/v1
    description: Production server
  - url: https://staging-api.sms-onboarding.com/v1
    description: Staging server
  - url: http://localhost:4000/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /integration/maintenance-portal/export:
    post:
      tags:
        - Integration
        - Export
      summary: Export vessel data to maintenance portal
      operationId: exportToMaintenance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vesselId
              properties:
                vesselId:
                  type: string
                  format: uuid
                includePhotos:
                  type: boolean
                  default: true
                format:
                  type: string
                  enum: [json, xml, csv]
                  default: json
      responses:
        '200':
          description: Export initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  exportId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [pending, processing, completed, failed]
                  estimatedCompletionTime:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /integration/maintenance-portal/sync-users:
    post:
      tags:
        - Integration
      summary: Sync users from maintenance portal
      operationId: syncUsers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                companyId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: User sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced:
                    type: integer
                  created:
                    type: integer
                  updated:
                    type: integer
                  failed:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    webhookAuth:
      type: apiKey
      in: header
      name: X-Webhook-Signature
...
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # Enums
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...